refactor: external-api 쪽 domain 패키지 이름을 infrastructure로 변경
