+++ matzip-app-external-api/src/main/java/com/woowacourse/matzip/application/RestaurantService.java
@@ -4,12 +4,12 @@ import com.woowacourse.matzip.application.response.RestaurantResponse;
 import com.woowacourse.matzip.application.response.RestaurantTitleResponse;
 import com.woowacourse.matzip.application.response.RestaurantTitlesResponse;
 import com.woowacourse.matzip.domain.restaurant.Restaurant;
-import com.woowacourse.matzip.domain.restaurant.RestaurantFindQueryFactory;
-import com.woowacourse.matzip.domain.restaurant.RestaurantQueryRepository;
 import com.woowacourse.matzip.domain.restaurant.RestaurantRepository;
 import com.woowacourse.matzip.domain.restaurant.SortCondition;
 import com.woowacourse.matzip.domain.review.ReviewRepository;
 import com.woowacourse.matzip.exception.RestaurantNotFoundException;
+import com.woowacourse.matzip.infrastructure.restaurant.RestaurantFindQueryFactory;
+import com.woowacourse.matzip.infrastructure.restaurant.RestaurantQueryRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.data.domain.PageRequest;
